requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
backcall==0.2.0
beautifulsoup4==4.9.3
boto==2.48.0
boto3==1.4.7
botocore==1.7.32
certifi==2017.7.27.1
chardet==4.0.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-environ==0.4.5
django-extensions==3.1.0
django-storages==1.11.1
docutils==0.14
Faker==6.6.2
filelock==3.0.12
gitdb2==2.0.3
GitPython==2.1.7
gunicorn==19.7.1
httpie==2.4.0
idna==2.6
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.9.3
mysqlclient==2.0.3
olefile==0.44
parso==0.8.1
pbr==5.5.1
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.11.15
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycairo==1.20.0
pycodestyle==2.7.0
Pygments==2.7.3
PySocks==1.7.1
python-dateutil==2.6.1
pytz==2020.5
pyunitreport==0.1.4
requests==2.25.1
requests-toolbelt==0.9.1
s3transfer==0.1.11
selenium==3.141.0
six==1.11.0
smmap2==2.0.3
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.55.1
testresources==2.0.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
